refactor: split escrow lifecycle entrypoints
