+++ contracts/escrow/src/create_contract.rs
@@ -1,55 +1,148 @@
-use soroban_sdk::vec;
+use crate::{
+    ttl, Contract, ContractStatus, DataKey, Error, Escrow, EscrowArgs, EscrowClient, Milestone,
+    ReleaseAuthorization,
+};
+use soroban_sdk::{contractimpl, symbol_short, Address, Env, Symbol, Vec};
 
-use crate::ContractStatus;
+#[contractimpl]
+impl Escrow {
+    /// Creates a new escrow contract with the specified client, freelancer, and milestone amounts.
+    ///
+    /// # Arguments
+    /// * `env` - The contract environment
+    /// * `client` - The address of the client funding the contract
+    /// * `freelancer` - The address of the freelancer performing the work
+    /// * `arbiter` - Optional arbiter address for dispute resolution
+    /// * `milestones` - Vector of milestone amounts (in stroops)
+    /// * `release_authorization` - Authorization mode for milestone releases
+    ///
+    /// # Returns
+    /// The unique contract ID
+    ///
+    /// # Errors
+    /// * `InvalidParticipants` - If client and freelancer are the same address
+    /// * `EmptyMilestones` - If no milestones are provided
+    /// * `InvalidMilestoneAmount` - If any milestone amount is <= 0
+    /// * `MissingArbiter` - If arbiter is required but not provided
+    /// * `InvalidArbiter` - If arbiter is same as client or freelancer
+    /// * `ContractIdOverflow` - If the next id would exceed `u32::MAX`
+    /// * `ContractIdCollision` - If the allocated id slot is already occupied
+    pub fn create_contract(
+        env: Env,
+        client: Address,
+        freelancer: Address,
+        arbiter: Option<Address>,
+        milestones: Vec<i128>,
+        release_authorization: ReleaseAuthorization,
+    ) -> u32 {
+        client.require_auth();
 
-use super::{assert_contract_state, create_client, setup};
+        if client == freelancer {
+            env.panic_with_error(Error::InvalidParticipant);
+        }
 
-#[test]
-fn creates_contract_and_persists_milestones() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let milestones = vec![&env, 200_0000000_i128, 400_0000000_i128, 600_0000000_i128];
+        match release_authorization {
+            ReleaseAuthorization::ArbiterOnly | ReleaseAuthorization::ClientAndArbiter
+                if arbiter.is_none() =>
+            {
+                env.panic_with_error(Error::MissingArbiter);
+            }
+            _ => {}
+        }
 
-    let contract_id = client.create_contract(&client_addr, &freelancer_addr, &milestones);
+        if let Some(ref arb) = arbiter {
+            if arb == &client || arb == &freelancer {
+                env.panic_with_error(Error::InvalidArbiter);
+            }
+        }
 
-    assert_eq!(contract_id, 1);
+        if milestones.is_empty() {
+            env.panic_with_error(Error::EmptyMilestones);
+        }
 
-    let contract = client.get_contract(&contract_id);
-    assert_contract_state(contract, ContractStatus::Created, 0, 0, 0);
+        for amount in milestones.iter() {
+            if amount <= 0 {
+                env.panic_with_error(Error::InvalidMilestoneAmount);
+            }
+        }
 
-    let stored_milestones = client.get_milestones(&contract_id);
-    assert_eq!(stored_milestones.len(), 3);
-    assert_eq!(stored_milestones.get(0).unwrap().amount, 200_0000000_i128);
-    assert_eq!(stored_milestones.get(1).unwrap().amount, 400_0000000_i128);
-    assert_eq!(stored_milestones.get(2).unwrap().amount, 600_0000000_i128);
+        let id = next_contract_id(&env);
+
+        ttl::extend_next_contract_id_ttl(&env);
+
+        let id = next_contract_id(&env);
+
+        let freelancer_addr = freelancer.clone();
+        let contract = Contract {
+            client: client.clone(),
+            freelancer: freelancer.clone(),
+            arbiter,
+            status: ContractStatus::Created,
+            funded_amount: 0,
+            released_amount: 0,
+            refunded_amount: 0,
+            release_authorization,
+        };
+        env.storage()
+            .persistent()
+            .set(&DataKey::Contract(id), &contract);
+
+        let mut milestone_vec: Vec<Milestone> = Vec::new(&env);
+        for amount in milestones.iter() {
+            milestone_vec.push_back(Milestone {
+                amount,
+                funded_amount: 0,
+                released: false,
+                refunded: false,
+                work_evidence: None,
+                refunded_amount: 0,
+            });
+        }
+        let milestone_key = Symbol::new(&env, "milestones");
+        env.storage()
+            .persistent()
+            .set(&(DataKey::Contract(id), milestone_key), &milestone_vec);
+
+        env.storage()
+            .persistent()
+            .set(&DataKey::NextContractId, &(id + 1));
+
+        env.events().publish(
+            (symbol_short!("created"), id),
+            (client, freelancer_addr, env.ledger().timestamp()),
+        );
+
+        id
+    }
 }
 
-#[test]
-#[should_panic]
-fn rejects_empty_milestones() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
+/// Returns the next contract id after verifying the slot is unused.
+fn next_contract_id(env: &Env) -> u32 {
+    let id: u32 = env
+        .storage()
+        .persistent()
+        .get(&DataKey::NextContractId)
+        .unwrap_or(1);
 
-    let milestones = vec![&env];
-    client.create_contract(&client_addr, &freelancer_addr, &milestones);
+    if env
+        .storage()
+        .persistent()
+        .get::<_, Contract>(&DataKey::Contract(id))
+        .is_some()
+    {
+        env.panic_with_error(Error::ContractIdCollision);
+    }
+
+    id
 }
 
-#[test]
-#[should_panic]
-fn rejects_zero_amount_milestone() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-
-    let milestones = vec![&env, 0_i128];
-    client.create_contract(&client_addr, &freelancer_addr, &milestones);
-}
-
-#[test]
-#[should_panic]
-fn rejects_same_participants() {
-    let (env, client_addr, _) = setup();
-    let client = create_client(&env);
-
-    let milestones = vec![&env, 100_0000000_i128];
-    client.create_contract(&client_addr, &client_addr, &milestones);
+/// Advances [`DataKey::NextContractId`] after a contract is persisted.
+#[allow(dead_code)]
+fn bump_next_contract_id(env: &Env, id: u32) {
+    let next_id = id
+        .checked_add(1)
+        .unwrap_or_else(|| env.panic_with_error(Error::ContractIdOverflow));
+    env.storage()
+        .persistent()
+        .set(&DataKey::NextContractId, &next_id);
 }
