+++ contracts/escrow/src/refund.rs
@@ -1,122 +1,120 @@
-use soroban_sdk::vec;
-
-use super::{
-    assert_contract_state, assert_milestone_flags, create_client, create_default_contract, setup,
+use crate::{
+    ttl, Contract, ContractStatus, DataKey, Error, Escrow, EscrowArgs, EscrowClient, Milestone,
 };
-use crate::ContractStatus;
+use soroban_sdk::{contractimpl, Env, Symbol, Vec};
 
-#[test]
-fn refunds_selected_unreleased_milestones_and_preserves_remaining_balance() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+#[contractimpl]
+impl Escrow {
+    /// Refunds unreleased milestones back to the client.
+    ///
+    /// # Arguments
+    /// * `env` - The contract environment
+    /// * `contract_id` - The contract ID
+    /// * `milestone_indices` - Vector of milestone indices to refund
+    ///
+    /// # Returns
+    /// The total amount refunded
+    ///
+    /// # Errors
+    /// * `ContractNotFound` - If contract doesn't exist
+    /// * `EmptyRefundRequest` - If milestone_indices is empty
+    /// * `DuplicateMilestoneInRefund` - If the same milestone appears multiple times
+    /// * `IndexOutOfBounds` - If any milestone index is out of bounds
+    /// * `AlreadyReleased` - If any milestone was already released
+    /// * `AlreadyRefunded` - If any milestone was already refunded
+    /// * `InsufficientFunds` - If contract doesn't have enough balance to refund
+    pub fn refund_unreleased_milestones(
+        env: Env,
+        contract_id: u32,
+        milestone_indices: Vec<u32>,
+    ) -> i128 {
+        if milestone_indices.is_empty() {
+            env.panic_with_error(Error::EmptyRefundRequest);
+        }
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    assert!(client.release_milestone(&contract_id, &0));
+        for i in 0..milestone_indices.len() {
+            for j in (i + 1)..milestone_indices.len() {
+                if milestone_indices.get(i).unwrap() == milestone_indices.get(j).unwrap() {
+                    env.panic_with_error(Error::DuplicateMilestoneInRefund);
+                }
+            }
+        }
 
-    let refund_ids = vec![&env, 1_u32];
-    let refunded = client.refund_unreleased_milestones(&contract_id, &refund_ids);
-    assert_eq!(refunded, 400_0000000_i128);
+        let mut contract: Contract = env
+            .storage()
+            .persistent()
+            .get(&DataKey::Contract(contract_id))
+            .unwrap_or_else(|| env.panic_with_error(Error::ContractNotFound));
 
-    let contract = client.get_contract(&contract_id);
-    assert_contract_state(
-        contract,
-        ContractStatus::Funded,
-        1_200_0000000_i128,
-        200_0000000_i128,
-        400_0000000_i128,
-    );
-    assert_milestone_flags(client.get_milestones(&contract_id), 1, false, true);
-    assert_eq!(
-        client.get_refundable_balance(&contract_id),
-        600_0000000_i128
-    );
-}
+        ttl::extend_contract_ttl(&env, contract_id);
 
-#[test]
-fn marks_contract_refunded_when_all_unreleased_milestones_are_refunded() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+        Self::require_not_finalized(&env, contract_id);
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    let refund_ids = vec![&env, 0_u32, 1_u32, 2_u32];
-    let refunded = client.refund_unreleased_milestones(&contract_id, &refund_ids);
-    assert_eq!(refunded, 1_200_0000000_i128);
+        contract.client.require_auth();
 
-    let contract = client.get_contract(&contract_id);
-    assert_contract_state(
-        contract,
-        ContractStatus::Refunded,
-        1_200_0000000_i128,
-        0,
-        1_200_0000000_i128,
-    );
-    assert_eq!(client.get_refundable_balance(&contract_id), 0);
-}
+        let milestone_key = Symbol::new(&env, "milestones");
+        let mut milestones: Vec<Milestone> = env
+            .storage()
+            .persistent()
+            .get(&(DataKey::Contract(contract_id), milestone_key.clone()))
+            .unwrap();
 
-#[test]
-#[should_panic]
-fn rejects_empty_refund_request() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+        ttl::extend_milestone_ttl(&env, contract_id);
 
-    let refund_ids = vec![&env];
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
-}
+        let mut total_refund_amount: i128 = 0;
 
-#[test]
-#[should_panic]
-fn rejects_duplicate_milestones_in_single_refund() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+        for idx in milestone_indices.iter() {
+            if idx >= milestones.len() {
+                env.panic_with_error(Error::IndexOutOfBounds);
+            }
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    let refund_ids = vec![&env, 1_u32, 1_u32];
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
-}
+            let milestone = milestones.get(idx).unwrap();
 
-#[test]
-#[should_panic]
-fn rejects_refunding_released_milestone() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+            if milestone.released {
+                env.panic_with_error(Error::AlreadyReleased);
+            }
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    assert!(client.release_milestone(&contract_id, &0));
+            if milestone.refunded {
+                env.panic_with_error(Error::AlreadyRefunded);
+            }
 
-    let refund_ids = vec![&env, 0_u32];
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
-}
+            total_refund_amount += milestone.amount;
+        }
 
-#[test]
-#[should_panic]
-fn rejects_refunding_same_milestone_twice() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+        let available_balance =
+            contract.funded_amount - contract.released_amount - contract.refunded_amount;
+        if available_balance < total_refund_amount {
+            env.panic_with_error(Error::InsufficientFunds);
+        }
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    let refund_ids = vec![&env, 2_u32];
-    assert_eq!(
-        client.refund_unreleased_milestones(&contract_id, &refund_ids),
-        600_0000000_i128
-    );
+        for idx in milestone_indices.iter() {
+            let mut milestone = milestones.get(idx).unwrap();
+            milestone.refunded = true;
+            milestones.set(idx, milestone);
+        }
 
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
-}
+        contract.refunded_amount += total_refund_amount;
 
-#[test]
-#[should_panic]
-fn rejects_refund_when_balance_is_not_available() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr);
+        let all_refunded_or_released = milestones.iter().all(|m| m.released || m.refunded);
+        if all_refunded_or_released {
+            let all_refunded = milestones.iter().all(|m| m.refunded);
+            if all_refunded {
+                contract.status = ContractStatus::Refunded;
+            } else {
+                contract.status = ContractStatus::Completed;
+            }
+        }
 
-    assert!(client.deposit_funds(&contract_id, &200_0000000_i128));
-    let refund_ids = vec![&env, 1_u32];
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
+        env.storage().persistent().set(
+            &(DataKey::Contract(contract_id), milestone_key),
+            &milestones,
+        );
+        env.storage()
+            .persistent()
+            .set(&DataKey::Contract(contract_id), &contract);
+
+        ttl::extend_contract_and_milestones_ttl(&env, contract_id);
+
+        total_refund_amount
+    }
 }
