+++ contracts/escrow/src/release.rs
@@ -1,92 +1,182 @@
-use soroban_sdk::vec;
-
-use super::{
-    assert_contract_state, assert_milestone_flags, create_client, create_default_contract, setup,
+use crate::{
+    approvals, ttl, Contract, ContractStatus, DataKey, Error, Escrow, EscrowArgs, EscrowClient,
+    Milestone, ReleaseAuthorization,
 };
-use crate::ContractStatus;
+use soroban_sdk::{contractimpl, Address, Env, Symbol, Vec};
 
-#[test]
-fn releases_funded_milestones_and_completes_when_all_are_released() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr, &None);
+#[contractimpl]
+impl Escrow {
+    /// Releases a specific milestone, transferring funds to the freelancer.
+    ///
+    /// Requires valid, non-expired approvals based on the contract's ReleaseAuthorization mode.
+    ///
+    /// # Arguments
+    /// * `env` - The contract environment
+    /// * `contract_id` - The contract ID
+    /// * `caller` - The address of the caller (must be authorized)
+    /// * `milestone_index` - The index of the milestone to release
+    ///
+    /// # Returns
+    /// `true` if release was successful
+    ///
+    /// # Errors
+    /// * `ContractNotFound` - If contract doesn't exist
+    /// * `InvalidState` - If contract is not in Funded state
+    /// * `InvalidMilestone` - If milestone index is out of bounds
+    /// * `AlreadyReleased` - If milestone was already released
+    /// * `AlreadyRefunded` - If milestone was already refunded
+    /// * `InsufficientFunds` - If contract doesn't have enough funded balance
+    /// * `InsufficientApprovals` - If required approvals are missing
+    /// * `ApprovalExpired` - If approvals have expired
+    /// * `UnauthorizedRole` - If caller is not authorized to release
+    ///
+    /// # Security
+    /// - Requires valid approvals that haven't expired
+    /// - Approvals are cleared after successful release
+    /// - Fail-closed: missing or expired approvals prevent release
+    pub fn release_milestone(
+        env: Env,
+        contract_id: u32,
+        caller: Address,
+        milestone_index: u32,
+    ) -> bool {
+        caller.require_auth();
 
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
+        let mut contract: Contract = env
+            .storage()
+            .persistent()
+            .get(&DataKey::Contract(contract_id))
+            .unwrap_or_else(|| env.panic_with_error(Error::ContractNotFound));
 
-    assert!(client.release_milestone(&contract_id, &0, &client_addr));
-    let contract = client.get_contract(&contract_id);
-    assert_contract_state(
-        contract,
-        ContractStatus::Funded,
-        1_200_0000000_i128,
-        200_0000000_i128,
-        0,
-    );
-    assert_milestone_flags(client.get_milestones(&contract_id), 0, true, false);
-    assert_eq!(
-        client.get_refundable_balance(&contract_id),
-        1_000_0000000_i128
-    );
+        ttl::extend_contract_ttl(&env, contract_id);
 
-    assert!(client.release_milestone(&contract_id, &1, &client_addr));
-    assert!(client.release_milestone(&contract_id, &2, &client_addr));
+        Self::require_not_finalized(&env, contract_id);
 
-    let contract = client.get_contract(&contract_id);
-    assert_contract_state(
-        contract,
-        ContractStatus::Completed,
-        1_200_0000000_i128,
-        1_200_0000000_i128,
-        0,
-    );
-    assert_eq!(client.get_refundable_balance(&contract_id), 0);
+        if contract.status != ContractStatus::Funded {
+            env.panic_with_error(Error::InvalidState);
+        }
+
+        let is_client = caller == contract.client;
+        let is_freelancer = caller == contract.freelancer;
+        let is_arbiter = contract.arbiter.as_ref() == Some(&caller);
+
+        match contract.release_authorization {
+            ReleaseAuthorization::ClientOnly => {
+                if !is_client {
+                    env.panic_with_error(Error::UnauthorizedRole);
+                }
+            }
+            ReleaseAuthorization::ArbiterOnly => {
+                if !is_arbiter {
+                    env.panic_with_error(Error::UnauthorizedRole);
+                }
+            }
+            ReleaseAuthorization::ClientAndArbiter => {
+                if !is_client && !is_arbiter {
+                    env.panic_with_error(Error::UnauthorizedRole);
+                }
+            }
+            ReleaseAuthorization::MultiSig => {
+                if !is_client && !is_freelancer {
+                    env.panic_with_error(Error::UnauthorizedRole);
+                }
+            }
+        }
+
+        approvals::check_approvals(&env, &contract, contract_id, milestone_index)
+            .unwrap_or_else(|e| env.panic_with_error(e));
+
+        let milestone_key = Symbol::new(&env, "milestones");
+        let mut milestones: Vec<Milestone> = env
+            .storage()
+            .persistent()
+            .get(&(DataKey::Contract(contract_id), milestone_key.clone()))
+            .unwrap();
+
+        ttl::extend_milestone_ttl(&env, contract_id);
+
+        if milestone_index >= milestones.len() {
+            env.panic_with_error(Error::IndexOutOfBounds);
+        }
+
+        let mut milestone = milestones.get(milestone_index).unwrap().clone();
+
+        if milestone.released {
+            env.panic_with_error(Error::MilestoneAlreadyReleased);
+        }
+
+        if milestone.refunded {
+            env.panic_with_error(Error::AlreadyRefunded);
+        }
+
+        let available_balance =
+            contract.funded_amount - contract.released_amount - contract.refunded_amount;
+        if available_balance < milestone.amount {
+            env.panic_with_error(Error::InsufficientFunds);
+        }
+
+        let _release_amount = milestone.amount;
+        milestone.released = true;
+        milestones.set(milestone_index, milestone.clone());
+        contract.released_amount += milestone.amount;
+
+        if is_initialized(&env) {
+            let fee_bps = get_protocol_fee_bps(&env);
+            if fee_bps > 0 {
+                let fee = calculate_protocol_fee(milestone.amount, fee_bps);
+                let current_accumulated: i128 = env
+                    .storage()
+                    .persistent()
+                    .get(&DataKey::AccumulatedProtocolFees)
+                    .unwrap_or(0);
+                env.storage().persistent().set(
+                    &DataKey::AccumulatedProtocolFees,
+                    &(current_accumulated + fee),
+                );
+            }
+        }
+
+        approvals::clear_approvals(&env, contract_id, milestone_index);
+
+        let all_released = milestones.iter().all(|m| m.released || m.refunded);
+        if all_released {
+            contract.status = ContractStatus::Completed;
+        }
+
+        env.storage().persistent().set(
+            &(DataKey::Contract(contract_id), milestone_key),
+            &milestones,
+        );
+        env.storage()
+            .persistent()
+            .set(&DataKey::Contract(contract_id), &contract);
+
+        ttl::extend_contract_and_milestones_ttl(&env, contract_id);
+
+        true
+    }
 }
 
-#[test]
-#[should_panic]
-fn rejects_release_without_sufficient_balance() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr, &None);
-
-    assert!(client.deposit_funds(&contract_id, &100_0000000_i128));
-    client.release_milestone(&contract_id, &0, &client_addr);
+/// Returns true if the contract has been initialized.
+fn is_initialized(env: &Env) -> bool {
+    env.storage()
+        .persistent()
+        .get::<_, bool>(&DataKey::Initialized)
+        .unwrap_or(false)
 }
 
-#[test]
-#[should_panic]
-fn rejects_release_of_invalid_milestone() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr, &None);
-
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    client.release_milestone(&contract_id, &3, &client_addr);
+/// Returns the protocol fee in basis points.
+fn get_protocol_fee_bps(env: &Env) -> u32 {
+    env.storage()
+        .persistent()
+        .get::<_, u32>(&DataKey::ProtocolFeeBps)
+        .unwrap_or(0)
 }
 
-#[test]
-#[should_panic]
-fn rejects_releasing_refunded_milestone() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr, &None);
-
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    let refund_ids = vec![&env, 1_u32];
-    client.refund_unreleased_milestones(&contract_id, &refund_ids);
-
-    client.release_milestone(&contract_id, &1, &client_addr);
-}
-
-#[test]
-#[should_panic]
-fn rejects_releasing_same_milestone_twice() {
-    let (env, client_addr, freelancer_addr) = setup();
-    let client = create_client(&env);
-    let contract_id = create_default_contract(&env, &client, &client_addr, &freelancer_addr, &None);
-
-    assert!(client.deposit_funds(&contract_id, &1_200_0000000_i128));
-    assert!(client.release_milestone(&contract_id, &0, &client_addr));
-
-    client.release_milestone(&contract_id, &0, &client_addr);
+/// Calculates the protocol fee for a given amount.
+fn calculate_protocol_fee(amount: i128, fee_bps: u32) -> i128 {
+    if fee_bps == 0 {
+        return 0;
+    }
+    amount * fee_bps as i128 / 10_000
 }
